Extract cache management functions to dedicated class
