+++ MatterControl.csproj
@@ -72,6 +72,7 @@
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
     <Compile Include="AboutPage\AboutWindow.cs" />
+    <Compile Include="AboutPage\CacheDirectory.cs" />
     <Compile Include="AboutPage\CheckForUpdatesPage.cs" />
     <Compile Include="ConfigurationPage\CalibrationSettings\CalibrationControls.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\ProbeCalibrationWizard.cs" />
